requirements.txt
flask
flask-cors
python-dotenv
langchain-core
langchain-google-genai